### 2019-06-07 #9 @whedon
[PRE REVIEW]: ChipSort: a SIMD and cache-aware sorting module
**Submitting author:** @nlw0 (<a href="http://orcid.org/0000-0003-0017-5890">Nicolau Werneck</a>)
**Repository:** <a href="https://github.com/nlw0/chipsort-article" target ="_blank">https://github.com/nlw0/chipsort-article</a>

Thanks for submitting your paper to JuliaCon @nlw0. The JuliaCon editor @vchuravy, will work with you on this issue to find a reviewer for your submission before creating the main review issue.

@nlw0 if you have any suggestions for potential reviewers then please mention them here in this thread. In addition, [this list of people](http://bit.ly/2V9GSFq) have already agreed to review for JuliaCon and may be suitable for this submission.

The JuliaCon submission bot @whedon is here to help you find and assign reviewers and start the main review. To find out what @whedon can do for you type:
