### 2019-06-07 #9 @vchuravy
@whendon assign @eschnett as reviewer

The timeline is until JuliaCon so July 20th, but the earlier the better :)

We have a reviewer guideline at https://proceedings.juliacon.org/guide/reviewers/ and if you have questions both @matbesancon and I are happy to help.
