### 2019-06-07 #9 @eschnett
@nlw0 In section 3.3 of your manuscript you compare against C++. Please add which compiler version from which vendor you are using, and in particular which libstdc++ you are using (I assume the compiler is less relevant than the standard library).
