### 2019-06-07 #9 @eschnett
@nlw0 Your manuscript is technical in the sense that it assumes that the reader is familiar with SIMD operations and tis benefits and drawbacks. Since the manuscript is also about sorting, that might not be true. You could add a subsection in the beginning describing SIMD, which operations are fast, and which are slow. Alternatively, maybe there is a good SIMD introduction available (online?) to which you could point.
