### 2019-06-07 #9 @eschnett
@nlw0 When sorting, one criterion is always whether the sort is stable or not. You don't discuss this. I assume that your sorting implementation is unstable. What would it take to implement a stable sort?
