### 2019-06-07 #9 @eschnett
@nlw0 All examples you show are for 32-bit quantities. What changes for different sizes (e.g. 8 bit, 64 bit, or even 128 bit md5sums)? What would need to change to sort floating point values with their pesky `nan` values? Maybe you want to add a few comments on this in the implementation or benchmark sections.
