### 2019-06-07 #9 @eschnett
@nlw0 You refer to "metaprogramming" in the manuscript, but don't explain the term. You could add a pointer, maybe referring to the respective section in the Julia manual.
