### 2019-06-07 #9 @eschnett
@vchuravy @nlw0 Overall evaluation: **Minor modification request**. I like this manuscript very much, it is very well written, technically thorough, and easy to read for a non-export. I would accept the manuscript as-is (except for the missing information regarding the C++ implementation), but would like @nlw0 to at least comment on my remarks.
