### 2019-06-08 #9 @nlw0
Thanks a lot, @eschnett. These are all excellent remarks and I'll try to address them as well as I can.
