### 2019-06-08 #12 @whedon
[REVIEW]: ChipSort: a SIMD and cache-aware sorting module
**Submitting author:** @nlw0 (<a href="http://orcid.org/0000-0003-0017-5890">Nicolau Werneck</a>)
**Repository:** <a href="https://github.com/nlw0/chipsort-article" target ="_blank">https://github.com/nlw0/chipsort-article</a>
