### 2019-06-09 #12 @eschnett
@vchuravy I see that the reviewer checklist has quite a few items ("general checks", "functionality", "documentation") that only make sense for software. So far I am assuming that I am reviewing a paper, not a software. Please clarify
