### 2019-06-10 #12 @vchuravy
Right, the checklist is based on the JOSS criteria and we are still iterating on it.

Our intent is to both review the paper and the software, the paper is the central piece, but we believe that it has to be backed by software that upholds the best practices of open-source and reproducible science.

We probably need to extend the checklist with a stronger focus on the paper part, your review in #9 is spot-on for that.
