### 2019-06-10 #12 @eschnett
@nlw0 In view of @vchuravy's comments regarding the code associated with the paper, I suggest
- add a pointer to the code to the paper, e.g. as reference
- make the code easier to use by registering the package
- add a licence to the code so that others can use it
