+++ paper/paper.tex
@@ -214,7 +214,7 @@ Figure~\ref{fig:bench-cdf} displays the measured times for $2^{13}$ and $2^{18}$
 \end{figure*}
 
 \subsection{Other platforms}
-A final experiment was carried out with different programming languages in the same machine and with Julia in different machines, to provide more context. Table~\ref{tab:bench-cxx} displays statistics collected from Julia, Python and C++ programs. For the case of Python the measurements obtained by calling the function through PyCall were better than using the native Python shell, therefore the best value was kept. Results obtained for C++ with Julia inter-operation using {\tt Cxx.jl} were slightly worse than native, thus discarded. While Numpy was almost as fast as Julia, no reason could be found for the dismal performance of the C++ standard library.
+One last experiment was carried out with different programming languages in the same machine and with Julia in different machines, to provide more context. Table~\ref{tab:bench-cxx} displays statistics collected from Julia, Python and C++ programs. Versions utilized were Julia 1.3.0-DEV.431, Python 3.6.8 with Numpy 1.14.3 and GCC 8.3.0 with libstdc++ 6.0.25 on a Ubuntu 18.04 system. For the case of Python the measurements obtained by calling the function through PyCall were better than using the native Python shell, therefore the best value was kept. Results obtained for C++ with Julia inter-operation using {\tt Cxx.jl} were slightly worse than native, thus discarded. While Numpy was almost as fast as Julia, no reason could be found for the dismal performance of the C++ standard library.
 
 \begin{table}[h]
 \tabcolsep11pt
@@ -222,10 +222,10 @@ A final experiment was carried out with different programming languages in the s
 \begin{tabular}{|l|r|r|}\hline
 Method & 8k & 1M\\\hline
 \chipsort & 172.35 $\mu$s & 35.28 ms\\\hline
-Julia standard & 345.46 $\mu$s & 68.39 ms\\\hline
-Numpy via PyCall & 369.50 $\mu$s & 70.01 ms\\\hline
-C++ {\tt std::qsort} & 788.59 $\mu$s & 147.37 ms\\\hline
-C++ {\tt std::sort} & 1,207.23 $\mu$s & 224.71 ms\\\hline
+Julia standard & 337.03 $\mu$s & 68.39 ms\\\hline
+Numpy via PyCall & 354.55 $\mu$s & 70.01 ms\\\hline
+C++ {\tt std::qsort} & 770.93 $\mu$s & 143.48 ms\\\hline
+C++ {\tt std::sort} & 1,177.54 $\mu$s & 221.27 ms\\\hline
 \end{tabular}}
 \label{tab:bench-cxx}
 \end{table}
