+++ paper/paper.tex
@@ -23,7 +23,9 @@ Sorting has enjoyed an important place within computer programming topics for a 
 
 Many factors can influence the running time of a computer program, starting with algorithmic complexity and the base clock speed of the processor. The performance observed in real computers has been increasingly depending on processor features that are not taken into account by the simplest models, though. These features include cache memory and parallelism at the processor and instruction levels. Cache memory has become popular in consumer processors in the past couple of decades, and as long as memory access patterns exhibit temporal and spatial locality, it allows programs to attain a higher performance than would be possible with a simpler memory architecture~\cite[Chapter 3]{Drepper07whatevery}. Writing software with better support for parallelism has also become necessary to utilize the full potential offered by modern and future processors~\cite{wilson2018}.
 
-The interest in sorting and the great relevance of cache memory and parallelism to modern computing have naturally led to research on sorting algorithms that account for these factors. Processors offering SIMD (single instruction, multiple data) instructions and other forms of instruction-level parallelism started to become widely available in the early 2000s. One example is the introduction of SSE instructions in the x86 architecture~\cite[Chapter 1]{kusswurm18}. The current availability of processors with 512 bit registers underscores the necessity of being able to write software that utilizes vector operations, what can also benefit GPU programming.
+One particular form of parallelism that concerns the present work are SIMD (single instruction, multiple data) instructions~\cite{SIMDtut}. This is a form of computation where a same operation is applied to different independent memory elements simultaneously. These operations can be described as operating over {\em vectors}, where each {\em lane} contains the same data that would be involved in carrying out the same operations as a scalar version of the same program. SIMD parallelism is only suitable to specific problems, and its use may imply in memory access constraints and even clock speed reductions. There are still cases where it can bring great performance while being conveniently employed.
+
+The interest in sorting and the great relevance of cache memory and parallelism to modern computing have naturally led to research on sorting algorithms that account for these factors. Processors offering SIMD instructions and other forms of instruction-level parallelism started to become widely available in the early 2000s. One example is the introduction of SSE instructions in the x86 architecture~\cite[Chapter 1]{kusswurm18}. The current availability of processors with 512 bit registers underscores the necessity of being able to write software that utilizes vector operations, what can also benefit GPU programming.
 
 At the time SIMD CPUs became widely available there was already a trend to employ specialized hardware like GPUs for scientific applications~\cite{larsen2001fast,DBLP:conf/micro/ThompsonHO02}. Although SIMD was used mostly for multimedia applications at first~\cite{CHEN2006509,DBLP:journals/mam/SlingerlandS05}, some early alternative uses include 3D graphics~\cite{DBLP:conf/pcm/MaY02} and machine learning~\cite{DBLP:conf/europar/StreyB01}.
 
