+++ paper/ref.bib
@@ -444,3 +444,11 @@ keywords = "Shellsort, sorting"
   year={1997},
   publisher={Wiley Online Library}
 }
+
+
+@online{SIMDtut,
+  title={{Comb} sort, or {Dobosiewicz} sort},
+  author={Martín Knoblauch Revuelta},
+  url={http://www.cs.uu.nl/docs/vakken/magr/2017-2018/files/SIMD%20Tutorial.pdf},
+  year={2013}
+}
