### 2019-06-22 ffcce83
moar simd and metaprogramming
