+++ paper/paper.tex
@@ -23,11 +23,11 @@ Sorting has enjoyed an important place within computer programming topics for a 
 
 Many factors can influence the running time of a computer program, starting with algorithmic complexity and the base clock speed of the processor. The performance observed in real computers has been increasingly depending on processor features that are not taken into account by the simplest models, though. These features include cache memory and parallelism at the processor and instruction levels. Cache memory has become popular in consumer processors in the past couple of decades, and as long as memory access patterns exhibit temporal and spatial locality, it allows programs to attain a higher performance than would be possible with a simpler memory architecture~\cite[Chapter 3]{Drepper07whatevery}. Writing software with better support for parallelism has also become necessary to utilize the full potential offered by modern and future processors~\cite{wilson2018}.
 
-One particular form of parallelism that concerns the present work are SIMD (single instruction, multiple data) instructions~\cite{SIMDtut}. This is a form of computation where a same operation is applied to different independent memory elements simultaneously. These operations can be described as operating over {\em vectors}, where each {\em lane} contains the same data that would be involved in carrying out the same operations as a scalar version of the same program. SIMD parallelism is only suitable to specific problems, and its use may imply in memory access constraints and even clock speed reductions. There are still cases where it can bring great performance while being conveniently employed.
+One particular form of parallelism that concerns our work is SIMD (single instruction, multiple data) instructions, where the processor applies a same operation to multiple independent memory elements simultaneously. These operations are said to act on {\em vectors}, where each parallel {\em lane} carries out the scalar version of the program. SIMD parallelism is only suitable to specific problems. It may introduce data access constraints and overheads, and even clock speed reductions. SIMD performance can be affected negatively by branching and scattered memory access. It is mostly suited for calculating many replicas of a long arithmetic expression, preferably with little IO. It can still pay off in appropriate settings and, as will be demonstrated, sometimes the implementation can be simple too. More information about SIMD can be found in~\cite{SIMDtut}.
 
-The interest in sorting and the great relevance of cache memory and parallelism to modern computing have naturally led to research on sorting algorithms that account for these factors. Processors offering SIMD instructions and other forms of instruction-level parallelism started to become widely available in the early 2000s. One example is the introduction of SSE instructions in the x86 architecture~\cite[Chapter 1]{kusswurm18}. The current availability of processors with 512 bit registers underscores the necessity of being able to write software that utilizes vector operations, what can also benefit GPU programming.
+Processors offering SIMD instructions and other forms of instruction-level parallelism started to become widely available in the early 2000s. One example is the introduction of SSE instructions in the x86 architecture~\cite[Chapter 1]{kusswurm18}. The current availability of processors with 512 bit registers, along with modern GPUs, underscore the demand for software-building tools that support vector operations. One programming language feature relevant to that goal is polymorphism, which allows programmers to abstract over data types and vectorization. Another is {\em metaprogramming}, the ability of writing programs that manipulate or generate other programs~\cite{Abelson1996}\cite[Ch.25]{JuliaDocs}. Metaprogramming enables code generation for complex algorithms customized to problem parameters and platform resources.
 
[…]
