+++ paper/ref.bib
@@ -185,7 +185,7 @@ doi = {10.1007/978-1-4842-4063-2}
 
 
 @article{CHEN2006509,
-title = "Implementation of H.264 encoder and decoder on personal computers",
+title = "Implementation of {H.264} encoder and decoder on personal computers",
 journal = "Journal of Visual Communication and Image Representation",
 volume = "17",
 number = "2",
@@ -447,8 +447,34 @@ keywords = "Shellsort, sorting"
 
 
 @online{SIMDtut,
-  title={{Comb} sort, or {Dobosiewicz} sort},
-  author={Martín Knoblauch Revuelta},
+  title={Practical {SIMD} programming},
+  author={Jacco Bikker},
   url={http://www.cs.uu.nl/docs/vakken/magr/2017-2018/files/SIMD%20Tutorial.pdf},
-  year={2013}
+  year={2017}
 }
+
+
+@online{JuliaDocs,
+  title={Julia 1.1 Documentation},
+  author={{The Julia Project}},
+  url={https://docs.julialang.org/en/v1.1/manual/metaprogramming/},
+  year={2019}
+}
+
+
+@book{Abelson1996,
+  added-at = {2012-12-17T22:51:51.000+0100},
+  address = {Cambridge},
+  author = {Abelson, Harold and Sussman, Gerald Jay and {with~Julie~Sussman}},
+  biburl = {https://www.bibsonomy.org/bibtex/211f37613ed3a6705eca058670186d406/gron},
+  edition = {2nd},
+  interhash = {775b3d446d8c4e21ec98778a82c025c7},
+  intrahash = {11f37613ed3a6705eca058670186d406},
+  isbn = {0-262-01153-0},
+  keywords = {LexicalAddressing SICP},
+  publisher = {MIT Press/McGraw-Hill},
+  timestamp = {2012-12-17T22:51:51.000+0100},
+  title = {Structure and Interpretation of Computer Programs},
+  year = 1996
+}
+
