+++ paper/paper.tex
@@ -25,11 +25,9 @@ Many factors can influence the running time of a computer program, starting with
 
 One particular form of parallelism that concerns our work is SIMD (single instruction, multiple data) instructions, where the processor applies a same operation to multiple independent memory elements simultaneously. These operations are said to act on {\em vectors}, where each parallel {\em lane} carries out the scalar version of the program. SIMD parallelism is only suitable to specific problems. It may introduce data access constraints and overheads, and even clock speed reductions. SIMD performance can be affected negatively by branching and scattered memory access. It is mostly suited for calculating many replicas of a long arithmetic expression, preferably with little IO. It can still pay off in appropriate settings and, as will be demonstrated, sometimes the implementation can be simple too. More information about SIMD can be found in~\cite{SIMDtut}.
 
-Processors offering SIMD instructions and other forms of instruction-level parallelism started to become widely available in the early 2000s. One example is the introduction of SSE instructions in the x86 architecture~\cite[Chapter 1]{kusswurm18}. The current availability of processors with 512 bit registers, along with modern GPUs, underscore the demand for software-building tools that support vector operations. One programming language feature relevant to that goal is polymorphism, which allows programmers to abstract over data types and vectorization. Another is {\em metaprogramming}, the ability of writing programs that manipulate or generate other programs~\cite{Abelson1996}\cite[Ch.25]{JuliaDocs}. Metaprogramming enables code generation for complex algorithms customized to problem parameters and platform resources.
+Processors offering SIMD or other forms of instruction-level parallelism started to become widely available in the early 2000s. One example is the introduction of SSE instructions in the x86 architecture~\cite[Chapter 1]{kusswurm18}. The current availability of processors with 512 bit registers, along with modern GPUs, underscore the demand for software-building tools that support vector operations. One programming language feature relevant to that goal is polymorphism, allowing programmers to abstract over data types and vectors. Another is {\em metaprogramming}, the ability of writing programs that manipulate or generate other programs~\cite{Abelson1996}\cite[Ch.25]{JuliaDocs}. Metaprogramming enables code generation for complex algorithms customized to problem parameters and platform resources.
 
-The interest in sorting and the great relevance of cache memory and parallelism to modern computing eventually led to research on sorting algorithms that account for these factors, following research in other topics where SIMD is more clearly beneficial. At the time SIMD CPUs became widely available there was already a trend to employ specialized hardware like GPUs for scientific applications~\cite{larsen2001fast,DBLP:conf/micro/ThompsonHO02}. Although SIMD was used mostly for multimedia applications at first~\cite{CHEN2006509,DBLP:journals/mam/SlingerlandS05}, some early alternative uses include 3D graphics~\cite{DBLP:conf/pcm/MaY02} and machine learning~\cite{DBLP:conf/europar/StreyB01}.
-
-Following earlier work on parallel sorting in other architectures, the late 2000s saw the first practical demonstrations of sorting with modern consumer SIMD chips~\cite{DBLP:conf/IEEEpact/InoueMKN07,DBLP:journals/pvldb/ChhuganiNLMHCBKD08}. Most of the techniques utilized then are still employed in more recent works~\cite{DBLP:journals/pvldb/BalkesenATO13,DBLP:journals/pvldb/InoueT15}.
[…]
