+++ paper/ref.bib
@@ -229,17 +229,17 @@ abstract = "H.264 is an emerging video coding standard, which aims at compressin
   organization={ACM}
 }
 
+@Comment crossref  = {DBLP:conf/micro/2002},
+
 @inproceedings{DBLP:conf/micro/ThompsonHO02,
   author    = {Chris J. Thompson and
                Sahngyun Hahn and
                Mark Oskin},
   title     = {Using modern graphics architectures for general-purpose computing:
                a framework and analysis},
-  booktitle = {Proceedings of the 35th Annual International Symposium on Microarchitecture,
-               Istanbul, Turkey, November 18-22, 2002},
+  booktitle = {Proceedings of the 35th Annual International Symposium on Microarchitecture},
   pages     = {306--317},
   year      = {2002},
-  crossref  = {DBLP:conf/micro/2002},
   url       = {https://doi.org/10.1109/MICRO.2002.1176259},
   doi       = {10.1109/MICRO.2002.1176259},
   timestamp = {Wed, 28 Nov 2018 12:57:16 +0100},
@@ -298,15 +298,15 @@ abstract = "H.264 is an emerging video coding standard, which aims at compressin
   bibsource = {dblp computer science bibliography, https://dblp.org}
 }
 
+@Comment crossref  = {DBLP:conf/europar/2001},
+
 @inproceedings{DBLP:conf/europar/StreyB01,
   author    = {Alfred Strey and
                Martin Bange},
   title     = {Performance Analysis of Intel's {MMX} and {SSE:} {A} Case Study},
-  booktitle = {Euro-Par 2001: Parallel Processing, 7th International Euro-Par Conference
-               Manchester, {UK} August 28-31, 2001, Proceedings},
+  booktitle = {Euro-Par 2001: Parallel Processing, 7th International Euro-Par Conference},
   pages     = {142--147},
   year      = {2001},
-  crossref  = {DBLP:conf/europar/2001},
   url       = {https://doi.org/10.1007/3-540-44681-8\_21},
   doi       = {10.1007/3-540-44681-8\_21},
   timestamp = {Tue, 14 May 2019 10:00:46 +0200},
@@ -449,7 +449,7 @@ keywords = "Shellsort, sorting"
 @online{SIMDtut,
   title={Practical {SIMD} programming},
   author={Jacco Bikker},
-  url={http://www.cs.uu.nl/docs/vakken/magr/2017-2018/files/SIMD%20Tutorial.pdf},
+  url={http://www.cs.uu.nl/docs/vakken/magr/2017-2018/files/SIMD\%20Tutorial.pdf},
   year={2017}
 }
 
@@ -460,6 +460,28 @@ keywords = "Shellsort, sorting"
   url={https://docs.julialang.org/en/v1.1/manual/metaprogramming/},
   year={2019}
 }
+
+@online{llvmD62272,
+  title={{LLVM} diff {D62272}},
+  author={Sanjay Patel},
+  url={https://reviews.llvm.org/D62272},
+  year={2019}
+}
+
+@online{llvm3acc4236b871,
+  title={{LLVM} diff {3acc4236b871}},
+  author={Craig Topper},
+  url={https://reviews.llvm.org/rG3acc4236b8717bf493b81905194a10e71f526702},
+  year={2019}
+}
+
+@online{Julia31442,
+  title={Julia issue \#31442},
+  author={Matt Bauman and others},
+  url={https://github.com/JuliaLang/julia/issues/31442},
+  year={2019}
+}
+
 
 
 @book{Abelson1996,
@@ -477,4 +499,3 @@ keywords = "Shellsort, sorting"
   title = {Structure and Interpretation of Computer Programs},
   year = 1996
 }
-
