+++ paper/ref.bib
@@ -499,3 +499,12 @@ keywords = "Shellsort, sorting"
   title = {Structure and Interpretation of Computer Programs},
   year = 1996
 }
+
+@misc{nicolau_leal_werneck_2019_3228848,
+  author       = {Nicolau Leal Werneck},
+  title        = {nlw0/ChipSort.jl: First release},
+  month        = may,
+  year         = 2019,
+  doi          = {10.5281/zenodo.3228848},
+  url          = {https://doi.org/10.5281/zenodo.3228848}
+}
