### 2019-06-22 #12 @nlw0
I believe I have addressed every point from the review. Please let me know if there's anything missing.
