### 2019-06-23 #12 @eschnett
@nlw0 Regarding the comparison to C++: Can you confirm that you used `-O2` and `-march=native` when building the C++ code? (The performance difference is striking.)
