### 2019-06-24 #12 @eschnett
The citation of the Chipsort Github repository should contain either a URL or a DOI. Similarly for
- Julia issue #31442
- Ultra-sort
- Sortingnetworks.jl
- LLVM diff D62272
- Comb sort
- Julia 1.1 documentation
- LLVM diff 3acc4236b871
- Julia-Con keynote address (URL for Youtube?)
