### 2019-06-24 #12 @eschnett
Regarding DOIs: I have recently started using DOIs to take snaphots of codes used in publications. Zenodo.org offers a convenient Github integration. In addition to pointing to the Github repo, Zenodo also hosts a tarball of the code version which is referenced by the DOI. The Journal of Open Source Software actually requires this step. I don't think JuliaCon does, but I still want to suggest it.
