### 2019-06-24 #12 @nlw0
I did use Zenodo to create a link, I learned it from you at SIMD.jl!... You can see it on refs.bib at the bottom. Problem is the style doesn't seem to like urls, either as text or even just a link on the PDF. I'm not sure what to do.
