### 2019-06-24 #12 @eschnett
I see; authors are not supposed to change the bibtex style. Can you try adding a bibtex "note=", and maybe use "\url{...}" in the note? Notes usually show up verbatim.
