### 2019-06-24 #12 @nlw0
bibtex seems to work by adding a field like   `howpublished = "\url{https://doi.org/10.5281/zenodo.3228848}"`
