### 2019-06-24 26dcc39
Fixing urls in references
