+++ paper/ref.bib
@@ -74,6 +74,7 @@
         author = {Sophie Wilson},
         year = {2018},
 	note = {JuliaCon keynote address},
+        howpublished = "\url{https://www.youtube.com/watch?v=zX4ZNfvw1cw}",
         url = {https://www.youtube.com/watch?v=zX4ZNfvw1cw}
 }
 
@@ -93,17 +94,16 @@
                Toshio Nakatani},
   title     = {{AA-Sort}: {A} New Parallel Sorting Algorithm for Multi-Core {SIMD}
                Processors},
-  booktitle = {16th International Conference on Parallel Architectures and Compilation
-               Techniques {(PACT} 2007), Brasov, Romania, September 15-19, 2007},
+  booktitle = {PACT},
   pages     = {189--198},
   year      = {2007},
-  crossref  = {DBLP:conf/IEEEpact/2007},
   url       = {http://doi.ieeecomputersociety.org/10.1109/PACT.2007.12},
   doi       = {10.1109/PACT.2007.12},
   timestamp = {Mon, 15 Apr 2019 16:16:42 +0200},
   biburl    = {https://dblp.org/rec/bib/conf/IEEEpact/InoueMKN07},
   bibsource = {dblp computer science bibliography, https://dblp.org}
 }
+  @Comment crossref  = {DBLP:conf/IEEEpact/2007},
 @proceedings{DBLP:conf/IEEEpact/2007,
   title     = {16th International Conference on Parallel Architectures and Compilation
                Techniques {(PACT} 2007), Brasov, Romania, September 15-19, 2007},
@@ -184,6 +184,7 @@ doi = {10.1007/978-1-4842-4063-2}
 }
 
 
+@Comment note = "Introduction: Special Issue on emerging H.264/AVC video coding standard",
 @article{CHEN2006509,
 title = "Implementation of {H.264} encoder and decoder on personal computers",
 journal = "Journal of Visual Communication and Image Representation",
@@ -191,7 +192,6 @@ volume = "17",
 number = "2",
 pages = "509 - 532",
 year = "2006",
-note = "Introduction: Special Issue on emerging H.264/AVC video coding standard",
 issn = "1047-3203",
 doi = "https://doi.org/10.1016/j.jvcir.2005.05.004",
 url = "http://www.sciencedirect.com/science/article/pii/S1047320305000398",
@@ -263,16 +263,14 @@ abstract = "H.264 is an emerging video coding standard, which aims at compressin
   bibsource = {dblp computer science bibliography, https://dblp.org}
 }
 
+  @Comment crossref  = {DBLP:conf/pcm/2002},
 @inproceedings{DBLP:conf/pcm/MaY02,
   author    = {Wan{-}Chun Ma and
                Chia{-}Lin Yang},
   title     = {Using {Intel} Streaming {SIMD} Extensions for {3D} Geometry Processing},
-  booktitle = {Advances in Multimedia Information Processing - {PCM} 2002, Third
-               {IEEE} Pacific Rim Conference on Multimedia, Hsinchu, Taiwan, December
-               16-18, 2002, Proceedings},
+  booktitle = {Third {IEEE} Pacific Rim Conference on Multimedia},
   pages     = {1080--1087},
   year      = {2002},
-  crossref  = {DBLP:conf/pcm/2002},
   url       = {https://doi.org/10.1007/3-540-36228-2\_134},
   doi       = {10.1007/3-540-36228-2\_134},
   timestamp = {Tue, 14 May 2019 10:00:54 +0200},
[…]
