### 2019-06-24 #12 @nlw0
Included urls and fixed listing spacing
