### 2019-07-05 #12 @vchuravy
Thank you.

@arfon this is the first paper ready to be accepted, IIRC you mentioned that you wanted to do it to make sure the pipeline fully works?
