### 2019-07-08 #12 @arfon
> Archive being a deposit with Zenodo?

Right.
