### 2019-07-08 #12 @vchuravy
Ah! Yes we should do that. Is there a way to do that automatically or is this something the user needs to do?
