### 2019-07-09 #12 @eschnett
I believe the code is already on Zenodo: <https://doi.org/10.5281/zenodo.3254944>.

The paper lives in a separate repo.
