### 2019-07-11 #12 @arfon
Looks like the reference checker is failing on these comments. I've fixed the bibtex in https://github.com/nlw0/chipsort-article/pull/1
