### 2019-07-12 #12 @arfon
@vchuravy - I've left a comment on https://github.com/JuliaCon/proceedings-papers/pull/3 that it would be good to address before we register the DOI with Crossref.
