### 2020-06-08 #12 @vchuravy
@nlw0 can add the missing DOIs?
