+++ paper/ref.bib
@@ -1,5 +1,6 @@
 @article{bezanson2017julia,
   title={Julia: A fresh approach to numerical computing},
+  doi={https://doi.org/10.1137/141000671},
   author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
   journal={SIAM review},
   volume={59},
@@ -222,6 +223,7 @@ abstract = "H.264 is an emerging video coding standard, which aims at compressin
 
 @inproceedings{larsen2001fast,
   title={Fast matrix multiplies using graphics hardware},
+  doi={https://doi.org/10.1145/582034.582089},
   author={Larsen, E Scott and McAllister, David},
   booktitle={Proceedings of the 2001 ACM/IEEE conference on Supercomputing},
   pages={55--55},
@@ -367,6 +369,7 @@ abstract = "H.264 is an emerging video coding standard, which aims at compressin
 
 @article{dobosiewicz1980efficient,
   title={An efficient variation of bubble sort},
+  doi={https://doi.org/10.1016/0020-0190(80)90022-8 },
   author={Dobosiewicz, Wlodzimierz},
   journal={Information Processing Letters},
   volume={11},
@@ -437,6 +440,7 @@ keywords = "Shellsort, sorting"
 
 @article{musser1997introspective,
   title={Introspective sorting and selection algorithms},
+  doi={https://doi.org/10.1002/(SICI)1097-024X(199708)27:8%3C983::AID-SPE117%3E3.0.CO;2-%23},
   author={Musser, David R},
   journal={Software: Practice and Experience},
   volume={27},
