### 2020-06-08 #12 @nlw0
Sure, I pushed a patch already. I could not find a DOI for SICP, though, and the suggested one is actually from a review article.
