### 2020-07-26 #12 @arfon
@nlw0 - could you merge this PR please? https://github.com/nlw0/chipsort-article/pull/2
