### 2020-07-26 #12 @arfon
@nlw0 - this paper is very nearly ready. Could you look at fixing up those DOIs that Whedon is complaining about? For example, changing:

`- https://doi.org/10.1137/141000671` to `10.1137/141000671`
