+++ paper/ref.bib
@@ -1,6 +1,6 @@
 @article{bezanson2017julia,
   title={Julia: A fresh approach to numerical computing},
-  doi={https://doi.org/10.1137/141000671},
+  doi={10.1137/141000671},
   author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B},
   journal={SIAM review},
   volume={59},
@@ -193,7 +193,7 @@ number = "2",
 pages = "509 - 532",
 year = "2006",
 issn = "1047-3203",
-doi = "https://doi.org/10.1016/j.jvcir.2005.05.004",
+doi = "10.1016/j.jvcir.2005.05.004",
 url = "http://www.sciencedirect.com/science/article/pii/S1047320305000398",
 author = "Yen-Kuang Chen and Eric Q. Li and Xiaosong Zhou and Steven Ge",
 keywords = "H.264, Video codec, Multimedia, MMX/SSE Technologies, SIMD, Hyper-Threading Technology, Multi-threading",
@@ -222,7 +222,7 @@ abstract = "H.264 is an emerging video coding standard, which aims at compressin
 
 @inproceedings{larsen2001fast,
   title={Fast matrix multiplies using graphics hardware},
-  doi={https://doi.org/10.1145/582034.582089},
+  doi={10.1145/582034.582089},
   author={Larsen, E Scott and McAllister, David},
   booktitle={Proceedings of the 2001 ACM/IEEE conference on Supercomputing},
   pages={55--55},
@@ -363,7 +363,7 @@ abstract = "H.264 is an emerging video coding standard, which aims at compressin
 
 @article{dobosiewicz1980efficient,
   title={An efficient variation of bubble sort},
-  doi={https://doi.org/10.1016/0020-0190(80)90022-8 },
+  doi={10.1016/0020-0190(80)90022-8 },
   author={Dobosiewicz, Wlodzimierz},
   journal={Information Processing Letters},
   volume={11},
@@ -425,7 +425,7 @@ number = "1",
 pages = "37 - 43",
 year = "1987",
 issn = "0020-0190",
-doi = "https://doi.org/10.1016/0020-0190(87)90034-2",
+doi = "10.1016/0020-0190(87)90034-2",
 url = "http://www.sciencedirect.com/science/article/pii/0020019087900342",
 author = "Janet Incerpi and Robert Sedgewick",
 keywords = "Shellsort, sorting"
@@ -434,7 +434,7 @@ keywords = "Shellsort, sorting"
 
 @article{musser1997introspective,
   title={Introspective sorting and selection algorithms},
-  doi={https://doi.org/10.1002/(SICI)1097-024X(199708)27:8%3C983::AID-SPE117%3E3.0.CO;2-%23},
+  doi={10.1002/(SICI)1097-024X(199708)27:8%3C983::AID-SPE117%3E3.0.CO;2-%23},
   author={Musser, David R},
   journal={Software: Practice and Experience},
   volume={27},
