### 2020-07-26 #12 @nlw0
I think I fixed it. I already looked into the SICP DOI, see above...
