### 2020-07-26 baa721f
trying to make massed-up doi with special characters work
