+++ paper/ref.bib
@@ -434,7 +434,7 @@ keywords = "Shellsort, sorting"
 
 @article{musser1997introspective,
   title={Introspective sorting and selection algorithms},
-  doi={10.1002/(SICI)1097-024X(199708)27:8%3C983::AID-SPE117%3E3.0.CO;2-%23},
+  doi={10.1002/(SICI)1097-024X(199708)27:8<983::AID-SPE117>3.0.CO;2-#},
   author={Musser, David R},
   journal={Software: Practice and Experience},
   volume={27},
