### 2020-07-26 #12 @arfon
@vchuravy @matbesancon - this one went through fine with no manual processing on my part.

We need to update the proceedings website to make sure that the branding an content are correct but I think the tooling is mostly in good shape now.
